Introduce null values for scalar types
The null value is used as the imported value for an `out' parameter in a
C function binding.  Previously a literal expression that was generated
based on the SML type of the parameter, for example, 0, 0.0, #"\000",
provided the imported value.

The advantage of a null value is that the code generator does not need
knowledge of the SML representation of a basic types to create a value.
This allows an alias of a basic type to be made abstract without needing
changes to C function bindings.  For example, GLib.Pid.t, currently
defined as FFI.Int.t, could be changed to Posix.Process.pid.