Move abstract type Pid.t to GLib.Pid.t
The type GLib.Pid.t is a GIR alias in the GLib namespace so does not
appear when generating code from TYPELIB files using the GIRepository
library.  When the switch is made to generate code from GIR files,
GLib.Pid.t will be generated as an alias type, which will be supplanted
by this abstract type.