Correct code generation for scalar `out' parameters
Scalar `out' parameters with the annotation `caller-allocates' generated
a reference to the function withRefNewFun that does not exist for scalar
types.  For a scalar `out' parameter, the caller necessarily allocates
the memory because there are insufficient levels of indirection to allow
the callee to allocate memory.  Therefore, `caller-allocates' and
`callee-allocates' are ignored for scalar parameters.