Add support for multiple library versions to code generation
The GIRepository interface specified in G_I_REPOSITORY_REPOSITORY is
updated in order to allow multiple versions of the same namespace to be
loaded.

The implementation in GIRepositoryRepository based on GIR files is
updated to fully support multiple versions.  The girepository C library
cannot support multiple versions so the implementation based on TYPELIB
files is updated to support the updated interface but fails if any
attempt is made to load two versions of a library.

The interface G_I_REPOSITORY_REPOSITORY is also updated to include the
functions `getVersion` and `prependSearchPath`.

The code generation is updated for changes to G_I_REPOSITORY_REPOSITORY.
The code generation from TYPELIB files is updated to use Pango-1.30, in
line with code generation from GIR files.