Generate library `dylib` values in polyml.sml files
The code generated for Poly/ML expects the `dylib` value for a library
to be bound to `lib<namespace>`.  These values were defined in the file
'auto/polyml.sml'.  The name is the same for all versions of a namespace
so this would not allow multiple versions of the same library.  Instead,
a binding is generated at the top of the polyml.sml file for each
namespace using the library file name provided by GIRepository, if it
exists.