Align GIRepository.Repository.getSharedLibrary for GIR and TYPELIB
The shared library string can be a comma-separated list of library file
names.  For GIR files, this function returned only the first library in
the list but for TYPELIB files, the function returns the full string
because it uses g_irepository_get_shared_library internally.  For GIR
files, the function is updated to return the full string and code
generation extracts the first library file name.