Make FFI types used in callbacks portable
The type `id` in the signature CALLBACK is used as an FFI type for
dispatch and destroy callback functions via instances of the functor
Callback.  On the SML side, `id` was `Word32.word` but on the C side,
the type was `unsigned int`.  For i386 and x86_64 platforms these have
the same underlying integer type.  For portability, `Word32.word` has
been changed to `C_UInt.word` which guarantees the same underlying
integer type.