Restructure source code directories
In the repository directory:
  - SML source code is under src/sml
  - C source code is under src/mlton or src/polyml

In the installation directory:
  - lib/sml/<x> is a copy of src/sml/<x>

With multiple versions of a library, an application determines which
library version to use.  Consequently, mlton.mlb and polyml.sml files
are not provided to load all libraries.  src/sml/polyml.sml is still
present to provide a recursive version of `PolyML.use`.

The file giraffe.h is no longer generated with MLton to declare symbols
exported from SML.  It is not possible to generate this file with
multiple versions of a library.  The required extern declarations for
functions exported from SML are declared manually in H files for each
library.