Introduce configure script for installation
The configure script checks for dependencies and creates the file
config.mk that is used by Makefile to determine which libraries to build
and install for which SML compilers.  As a result, it is no longer
necessary for all supported libraries to be present and no error message
occurs when an SML compiler is not available.  The README and INSTALL
files are updated accordingly.