Add support for Darwin (Mac OS X)
This requires updates for Poly/ML.

Updated the build and installation process to:
  - use the DYLIB file extension for shared libraries
  - use an alternative linker command for dyld
if the kernel name is Darwin.

In the SML code, shared library file names are no longer specified as
arguments to `PolyMLFFI.load_lib`.  Mentioning these file names affects
the portability of the code because the shared library file names on
other platforms are not entirely predictable.  The empty string is given
to `PolyMLFFI.load_lib` which causes null to be passed to dlopen
internally.  As a result, applications must add library dependencies
to the binary by specifying additional -l options.  With GNU ld,
applications must ensure that --no-as-needed behaviour is used.  For an
interactive session, libraries should be specified using LD_PRELOAD.

Require Poly/ML 5.5.2 to allow:
 1. null to be passed to dlopen (see Poly/ML commit r1906);
 2. applications to specify additional linker flags which prevents use
    of polyc so `pkg-config polyml --libs` must be used instead.

For item 2, additional linker flags are required because:
 A. dyld does not support LD_RUN_PATH so the -rpath option is used;
 B. passing null to dlopen means that additional -l options must
    be specified to add library dependencies to the binary.