Remove fixes and extra files for TYPELIB code generation
The code generated from TYPELIB files will never be used as a library
because it is missing alias types and uses platform-specific int sizes.
The purpose of the code generated from TYPELIB files is for comparison
with the code generated from GIR files with to find code generation
errors.  This comparison can be done on the raw output.  Therefore the
fixes and extra files are not required for TYPELIB code generation.