Moved codegen source to codegen/src/
codegen-typelib.sml and codegen-gir.sml are moved to codegen/src/ and
split into multiple files in codegen/src/codegen.  codegen-typelib.sml
is split into codegen-typelib-${N}.sml for N in 1, 2 to have multiple
sessions to support different versions of the same namespace.

The script called "codegen" is introduced to invoke code generation and
apply code fixes, and copy the outputs into the working directory.  This
script incorporates the existing scripts make-gir and copy-gir which are
removed.

codegen/README is updated to document the code generation process.