Reimplement nested 'use' function with PolyML.getUseFileName
This change should have no effect on the observable behaviour of 'use'
for the simple scripts in the "polyml.sml" files.  It is cleaner to
prefix relative file paths with the directory of the file that uses them
rather than changing the current directory state to the location of each
file as it is used.

PolyML.getUseFileName was introduced in Poly/ML 5.6.