Update and extend C pointer abstraction
The signature POINTER, which represents an untyped pointer, is renamed
to C_POINTER and is extended by a new signature C_TYPED_POINTER that
represents a pointer to a specific C type.  A functor CTypedPointer is
introduced to created a typed pointer given a C_TYPE that characterizes
the type of data pointed to.  Typed pointers are used to simplify the
implementations of C_ARRAY.

Null type-safety

For each compiler, the implementation of C_POINTER required backdoors to
the underlying pointer type to break null-type-safety because a pointer
passed by reference may have different nullity for the before and after
value.  Values that are passed by reference are held in mutable state so
there is a conflict in preserving null-type safety through mutable state
and passing pointers by reference.  Given this, null-type-safety is
clarified here.  Specifically, null-type safety is not preserved through
mutable state and, therefore, a reference to a pointer, represented by
the type ('a, 'b) r.  The nullity of a pointer stored in a reference of
type ('a, 'b) r must match 'a, however, the nullity of the pointer read
from the reference is assumed to be 'b.