Restructure array FFI modules to support all array types
The functor FFIArray creates an implementation of F_F_I_ARRAY given an
implementation of C_ARRAY.  Functors are provided to create
implementations of C_ARRAY as follows:
  CVector   - for zero-terminated arrays
  CVectorN  - for length-specific arrays
These functors take an implementation of C_ARRAY_TYPE that provides
underlying array operations.  Functors are provided to create
implementations of C_ARRAY_TYPE as follows:
  CVectorType         - for zero-terminated arrays of scalars
  CVectorVectorType   - for zero-terminated arrays of arrays
  CVectorVectorNType  - for length-specific arrays of arrays
For scalar elements, the functors take an implementation of C_TYPE that
provides operations on the underlying C type.  For array elements the
functors take an implementation of C_ARRAY_TYPE, itself created by one
of the above functors.