Define record types using common signatures and functors
The signatures RECORD and NEW_RECORD are introduced to describe the
interface that is common to all record types.  NEW_RECORD extends RECORD
with an additional function C.withNewPtr that allocates a new instance
of a record and is provided for 'out' parameters where the caller must
allocate memory.

The functors BoxedRecord and PointerRecord are introduced to create
structures that implement RECORD.  NewBoxedRecord is introduced to
create structures that implement NEW_RECORD.  PointerRecord is for
record types that are internally just a pointer, also known as disguised
structures.