Change local class type name to `class`
For a class C, in structures <C> and <C>Class, the local type `'a class`
is the class type and the local type `t` is the base class type.  The
correspondence between old types and new types is as follows:

    'a <C>Class.t   --->  'a <C>Class.class
    <none>          --->     <C>Class.t  (new local type)
    'a <C>.class_t  --->  'a <C>.class
       <C>.t        <-->     <C>.t  (no change)

The rationale for this change is to ensure that `<C>Class.t` has no type
parameters in order to provide a similar interface to a struct type S
where `<S>Record.t` has no type parameter.

Although union types are not supported by the codegen, the hand code for
GdkEvent is updated to introduce the type `'a GdkEvent.union` for the
union type and `GdkEvent.t` is now the base union type and has no type
parameters.