Use FFI types and functions of non-root classes
Previously, for all classes descended from GObject, it was necessary to
use types and functions from `GObjectObjectClass.C` for 'in' parameters
because descendent classes did not provide their own 'with' functions.
Now that every class structure implements the signature CLASS, codegen
uses types and functions from that class structure.

For example, for a GtkWidget 'in' parameter, codegen would now use
`GtkWidgetClass.C.withPtr` instead of `GObjectObjectClass.C.withPtr`.

As a result of this change, the low-level and high-level FFIs now
enforce a class that is, or is a descendent of, the required type.
However, this does not change the existing interface because the
signature constraint already provided the required class constraint.