Introduce common interface for C types and support for C arrays
The purpose of a common interface for C types is to enable an interface
to be constructed for a C array type by functor application, given an
interface for the C type of the array elements.

The following signatures are introduced to specify interfaces for C types
as follows:
  - C_SCALAR      - C scalar type
  - C_ARRAY       - C array type
  - CLASS         - GObject object or interface
  - RECORD        - GObject struct
  - VALUE_RECORD  - GObject struct that can be used as a C value
  - UNION         - GObject union
  - ENUM          - GObject enum
  - FLAGS         - GObject bit flags

An interface signature contains a sub-structure FFI that provides
high-level FFI types and functions:

    structure FFI :
      sig
        (* if C value type *)
        type v
        ...
        val withVal : (v -> 'a) -> t -> 'a
        ...

        (* if C reference type *)
        type 'a p
        type notnull
        ...
        val withPtr : (notnull p -> 'a) -> t -> 'a
        ...
      end

and also a sub-structure C that provides underlying C representations
and operations:

    structure C :
      sig
        structure ValueType : C_VALUE_[EQ_][NULL_]TYPE  (* optional *)
        structure PointerType : C_POINTER_TYPE          (* optional *)
        structure ArrayType : C_ARRAY_TYPE            (* array only *)
      end

All C types have an implementation of at least one of the following
C type structures:
  - C_VALUE_[EQ_][NULL_]TYPE
  - C_POINTER_TYPE
  - C_ARRAY_TYPE
C types that can be used as C values (assigned to, for example) have an
implementation of C_VALUE_[EQ_][NULL_]TYPE.  C types that are references
(their C value is a pointer) have an implementation of C_POINTER_TYPE.
Only C array types have an implementation of C_ARRAY_TYPE.  The
signature C_ARRAY_TYPE includes C_POINTER_TYPE if 'a from_p = 'a.

Implementations of C_VALUE_[EQ_][NULL_]TYPE are provided for GObject
numeric C types as follows:
  - GBoolType
  - GCharType
  - GShortType
  - GUShortType
  - GIntType
  - GUIntType
  - GLongType
  - GULongType
  - GInt8Type
  - GUInt8Type
  - GInt16Type
  - GUInt16Type
  - GInt32Type
  - GUInt32Type
  - GInt64Type
  - GUInt64Type
  - GSSizeType
  - GSizeType
  - GFloatType
  - GDoubleType

Implementations of C_ARRAY_TYPE are constructed using functors as
follows:
  CValueCVectorType     - for zero-terminated arrays
    given an implementation of C_VALUE_NULL_TYPE
  CValueCVectorNType    - for length-specific arrays
    given an implementation of C_VALUE_TYPE
  CPointerCVectorType   - for zero-terminated arrays
  CPointerCVectorNType  - for length-specific arrays
    given an implementation of C_POINTER_TYPE

An implementation of C_SCALAR is constructed using the functor CScalar,
given an implementation of C_VALUE_[EQ_][NULL_]TYPE.  An implementation
of C_ARRAY is constructed using the functors
  CVector   - for zero-terminated arrays
  CVectorN  - for length-specific arrays
    given an implementation of C_ARRAY_TYPE.

Implementations of C_SCALAR are provided for GObject numeric C types as
follows:
  - GBool
  - GChar
  - GShort
  - GUShort
  - GInt
  - GUInt
  - GLong
  - GULong
  - GInt8
  - GUInt8
  - GInt16
  - GUInt16
  - GInt32
  - GUInt32
  - GInt64
  - GUInt64
  - GSSize
  - GSize
  - GFloat
  - GDouble

Implementations of CLASS, ENUM and FLAGS are produced by codgen for GObject
types.  Implementations of RECORD, VALUE_RECORD and UNION must still be
produced manually.

Support is added to codegen for:
  - function parameters and return value of a C array type
  - signal parameters of a C array type

Support is not present in codegen for:
  - signal return value of a C array type
  - properties of a C array type