Update codegen to use a tuple for function arguments
For functions that are methods, the object/interface argument is the
first curried argument and any remaining arguments are in the second
curried argument, which may be a tuple.

For signal handlers, all arguments are passed as a single tuple.   Note
that signal handlers do not include an argument for the object/interface
instance on which the signal was emitted.