Fix checks on GIR type information in codegen
The checks on an array element type were incorrectly rejecting a struct
that was not a reference.

The consistency checks on c:type attributes were rejecting types due to
invalid C type information, in particular for nested types.  This is
resolved by ignoring c:type attributes for nested types.  For top-level
types, the inconsistency appears to be due to incorrect introspection
annotations.