Use non-option type for 'out' parameters conditional on error
'out' parameters for functions with a boolean return type are assumed to
be written only if the return value is not false.  Therefore an option
type is used for the 'out' parameters to ensure only initialized values
can be used by the application.  However, for functions with a boolean
return type that raise an error use the return value to indicate whether
an error was raised.  In this case, 'out' parameters should not use an
option type because an exception is raised in SML when an error occurs
so uninitialized values cannot be used by the application anyway.  Also
the boolean return value should be ignored.  Therefore the SML return
type is `unit`.