Load GLib namespace before GObject
GLib modules depend on the ValueAccessor module to provides GValue
accessors which in turn depends on GObject.Type and GObject.ValueRecord.
To support this, the GLib namespace directory includes an initial
version of the GObject namespace structure that provides only
GObject.Type and GObject.ValueRecord and a module ValueAccessor that
provides access to GValues.  (ValueAccessor provides what was previously
in GObject.Value.)  Note that this does not force applications using
GLib to depend on libgobject: a program would depend on libgobject only
if it actually used a function provided by libgobject which would occur
only by using GObject or ValueAccessor.

Support is added to code generation for creating the initial GObject
namespace module in GLib.  This support allows initial namespace modules
to be generated in the directory init-<namespace>-<version>.  Subsequent
code generation of the directory <namespace>-<version> will rename and
extend init-<namespace>-<version> if such a directory exists.  This
support may also be useful for other namespaces if they have mutual
dependencies.