Do not generate getType functions for internal GTypes
Internal GTypes are those whose getTypeSymbol is given as "intern"
because there is no ..._get_type C function available to return a GType.
It is important that a getType function is not generated if there is no
..._get_type C function because a C function with that name could be
introduced for another purpose, e.g. g_variant_get_type.

GObject.Type is updated to include all fundamental GTypes.