Check that array length parameters are compatible with their arrays
This requires that an array parameter and its length parameter have the
same direction and that an array return value has a length parameter
with direction 'out'.  There are examples of functions where this is not
the case, for example g_io_channel_read_chars, but such functions are
currently unsupported because the information provided by introspection
annotations is not sufficiently understood in those cases.