Fix codegen for scalar alias 'out' parameters
In a high-level function, the argument given to the low-level function
for a scalar 'out' parameter is the null value of the C scalar interface
structure.  For alias types, this was previously using the interface for
the underlying C type, for example GInt.null, instead of the interface
for the alias type, for example GLib.Pid.null.