Fix codegen for arrays of value-like structs
Previously, it was assumed that elements of a struct array were
references, i.e. pointers to the struct, for all struct types.  For
an array of value-like structs, it should be assumed that the structs
are packed as values.

(The configuration parameter structTypes in config.sml determines
whether a struct is value-like.)