Update codegen to use attribute 'nullable' in GIR files
In the girepository library for GIR files, the functions
  - GIRepository.ArgInfo.mayBeNull
  - GIRepository.CallableInfo.mayReturnNull
are updated to use the attribute 'nullable' that indicates whether a
parameter or return value may pass a value that is null, making the
type of the value optional.  Furthermore, the function
  - GIRepository.ArgInfo.isOptional
is updated to use the attribute 'optional' that indicates whether an
'out' parameter can be ignored.  This function is not currently used by
codegen because an 'out' parameter is never ignored.

The changes that added support for attributes 'nullable' and 'optional'
were introduced in gobject-introspection 1.41.  For details, see

  https://bugzilla.gnome.org/show_bug.cgi?id=660879

GIR files generated by older versions of gobject-introspection are no
longer supported because the attribute 'allow-none' is ignored.