Use GType of derived class, not root class, in class GValue accessors
Previously, for a class X derived from GObject, `<X>Class.t` used the
GType for GObject, the root class.  `ValueAccessor.baseType <X>Class.t`
returned the GType for GObject.  With this change, `<X>Class.t` has the
GType for class X.  In the structure ValueAccessor, the function
`baseType` is renamed to `gtype`.