Generate correct property operators for mode and add 'new' operator
Previously, a construct-only property had a 'set' operator if writable
but should not.  A 'new' operator is added for writable properties to
allow a value to be provided at object construction.

Property types to represent modes are removed: the specification of a
property in a signature makes the record type explicit to show which
of the 'get', 'set' and 'new' operators are available for the property.