Replace GObject.Object.new with manual type-safe version
The object type is specified using a GValue accessor which provides both
the static type and the dynamic type (GType) of the new object.  Initial
values for properties are specified by property values created using the
'new' property operator.  Static types ensure that only a property of
the object class (or ancestor class) can be specified and that the value
specified for a property has the property type.