Introduce common interface for callback types
The signatures CALLBACK and NOTIFIED_CALLBACK are introduced to describe
the interface that is common to all callback types.  CALLBACK is used
for callbacks with scope 'async' and 'call'.  NOTIFIED_CALLBACK is used
for callbacks with scope 'notified'.

C function wrappers are no longer needed to fill in dispatch and destroy
arguments.  The callback interface provides high-level FFI functions for
dispatch and destroy parameters.