Provide C code dependencies for value-like record types
The implementation of the C interface for value-like record types
provided by the functor BoxedValueRecord is no longer parameterized by
functions to allocate (*_new) and free (*_free) struct values.  These
are not needed given that it is parameterized by a function to provide
the size of the struct (*_size).  Therefore, C functions to allocate and
free struct values are no longer required.

In order to support value-like record types that own other resources,
e.g. a heap-allocated string, a 'deep copy' function is required for
copying and a 'clear' function is required to release owned resourced
before freeing.  There is no mechanism to determine these automatically
so the constructor ValueRecord is parameterized, optionally, by C names
for these functions to allow them to be specified by configuration.
Therefore, the functor BoxedValueRecord is parameterized by a copy
function and a clear function which are calls to the C function names if
specfied and, otherwise, calls to memcpy (using the size of the struct)
and `ignore`, respectively.  Examples of value-like record types that
own other resources are GValue and GtkTargetEntry.

For code generation from GIR files, which produces portable code,
C code that is compiled on the target platform is required to provide
a function returning the size of a struct.  Support is added for
generating C code as described below.  For code generation from TYPELIB
files, which does not need to produce portable code, no C code is
required because the function to return the size of the a struct can
hardcode the value from the TYPELIB file.

The generated C source code is output in each namespace directory so it
no longer makes sense for namespace directories to be stored under
src/sml/ because they no longer contain only SML code.  The directory
src is restructured in that src/sml/* is moved into src/.  (This also
simplifies the generation of all C code in future because the generated
output directory corresponds to src/ rather than src/sml/ so it is
possible to generate code in, say, src/mlton without restructuring the
existing codegen output directory.)

To enable C code to be generated using the GIRepository interface for
GIR files, the function `GIRepository.StructInfo.getCName` is added
which returns the C struct type name as a string and the function
`GIRepository.BaseInfo.getVersion` is added to return the version that
introduced the library entity.