Change file naming convention to avoid conflicts
The current approach uses the module name with the extension '.sml' as
the file name.  This does not give a one-to-one map from module names to
file names for two reasons:

 1. The namespaces for signatures, structures and functors are separate
    so it is possible, although not good practice, for distinct modules
    to have the same name but these would not have distinct file names.

 2. The file system may have case-insensitive file names so even if
    module names are distinct, the file names may not be.

With the current module names, signatures, structures and functors have
different names so item 1 is not currently an issue.  Item 2 is an issue
because some module names differ only by case, for example, signature
FLAGS and structure Flags.  FAT32 has case-insensitive file names and
HFS+ may be configured to have case-insensitive file names.

The new file naming convention suffixes the module name with '-sig'
for a signature and '-fun' for a functor.  This avoids the issue for
item 1 and mostly for item 2.  It does not give distinct file names for
modules of the same type whose names differ only by case but this is
unlikely to occur in practice.