Use abstract SML type in C array interface
The SML type used for C function arguments and return values that have
an array type is changed from a list or vector type, that is internally
converted to and from a C array, to an abstract type that represents a
C array itself.  This allows a C array returned by one C function to be
passed to another C function without the overhead of conversion to and
from an SML list or vector type.  This does not apply to the array type
for a string, i.e. const gchar *, because requiring application code to
explicitly convert between string literals and the abstract SML type for
the C array would be too inconvenient.

The abstract SML type for a C array is the type t in the interface
structure for a C array that is created using functor CArray or CArrayN.
The correspondence between old and new functors is as follows:

    CVectorLowLevel      --->  CArray
    CVectorNLowLevel     --->  CArrayN
    CVector              --->  VectorCArray
    CVectorN             --->  VectorCArrayN

In other functor names, 'CVector' is changed to 'CArray':

    CPointerCVectorNType ---> CPointerCArrayNType
    CPointerCVectorType  ---> CPointerCArrayType
    CValueCVectorNType   ---> CValueCArrayNType
    CValueCVectorType    ---> CValueCArrayType

Previously the interface structures for a C array was declared locally
in each namespace module that required it.  Now these structures
must be declared globally because they appear in the function signatures
which presents some issues:

 1. As the type t in these C array structures is abstract and and
    functor application is generative, only a single instance of the
    interface structure for each C array type can be evaluated.  It
    is not possible to declare all possible structures up front
    because there are an infinite number when we consider arrays of
    arrays.  Therefore an interface structure must be declared only
    if required.  However if an interface structure is not required
    by the namespace that introduces its element type, it could
    still be required by a later namespace.  A later namespace
    cannot evaluate its own instance of the structure because there
    may be more than one such namespace.  Therefore it must be
    possible to evaluate a single instance of the structure if it is
    not required by the namespace that declares its element type.

 2. The interface structures are referenced in types in signatures.
    If an interface structure is required by the namespace that
    declares its element type then it is not possible to reference
    the structure because it hasn't yet been declared.  Therefore,
    the interface structure must also be added to the namespace.

The approach taken is described below, with arrays generalized to
containers.  However, only C arrays are currently supported.

An interface structure is declared for each container type, if it is
required, whose name has the form

    <ElementType><ContainerType>

where ContainerType is

    CArray
      for a zero-terminated C array

    CArrayN
      for a C array

The container interface structure is always added in the directory for
namespace of the element type except for fundamental types, e.g. gint,
which are introduced in the GLib namespace.  (These containers must be
introduced in GLib in the scope of the initial GObject namespace to
ensure that the container structures are extended with value accessors.
This doesn't apply to Utf8 because GType provides the string accessor.)

 1. For a container type that used in a namespace other than the
    namespace that introduces the element type, the structure is not
    added to the namespace of the element type.

 2. For a container type used in the namespace that introduces the
    element types, it is necessary to include the container type in
    the namespace module.

Therefore, global references to the container interface structure cannot
be used because the interface structure may not be in the namespace
structure that declares its element type, e.g.

    GObjectObjectClassCArrayN

must be used instead of

    GObject.ObjectClassCArrayN

(An existing namespace structure cannot be extended because that does
not work generally with MLB files.)  A container interface structure
must be included only once due to the generative nature of functors.
For Poly/ML, the file for each container structure is included by a
variant of the use command that checks whether a structure already
exists.  For MLton, the file for each container structure is included
via its own MLB file.  The semantics of MLB files ensure that their
contents are evaluated only once.