Use SML type int for C types gsize and gssize
Although the range of the SML type int is not usually large enough to
contain the ranges of the C types gsize and gssize, the SML type int is
typically used for sizes in SML and aligning these types helps to avoid
conversion between SML integer types in application code where Basis
Library functions are mixed with C functions.

The representation of the C type GType is a pointer internally and GType
is defined as gsize, which is guaranteed to be wide enough to hold a
pointer.  The SML type int is unlikely to be wide enough to hold a
pointer, especially on 64 bit platforms so we do not want to inherit the
int representation of gsize for GType.  For now, the C interface for
GType, which is abstract, is changed to use the C interface for gulong
internally, which uses LargeInt.int on the SML side.  On some platforms
it is possibly that LargeInt.int is not large enough to hold a pointer.
However, this is no worse than before because the C interface for gsize
used to be defined as the C interface for gulong internally.

Support is also added for an array length parameter with C type guint8.