Ensure unique interface structure names for C array types
The name of an interface structure for a C array type is updated to use
a form that distinguishes whether the elements are pointers.  This
avoids the potential for a name clash where the values of a C element
type can be contained in an array either packed inline or referenced by
a pointer.  In principle, the name clash could occur for a C element
type that is a struct type but GObject Introspection does not currently
provide sufficient information to distinguish the two types of C array.
(In GObject-based libraries, it is expected that one form of C array is
used consistently for a given element type.)

The interface structure name for a C container type has the form

    <ElementType><ContainerType>

For a C array, ContainerType was

    CArray<N>

but is now

    C<Ptr>Array<N>

where Ptr is

    Ptr
      if the C array elements are pointers to the element values

    <>
      if the C array elements are the element values packed inline