Factor out C memory operations from FFI framework
These changes pave the way to move FFI modules that are GLib-specific
out of the FFI directory.

Prior to this commit, there were several FFI modules with bindings to
g_malloc and g_free.  To avoid these bindings in FFI modules, the
signature C_MEMORY is introduced which provides the required memory
model.  Then, structures that allocate and free memory are changed to
functors that are parameterized by C_MEMORY.

This affects the structures that implement C_VALUE_[EQ_][NULL_]TYPE for
scalar types and CPointerInternal.  Also, CPointerInternal, which
exposes the underlying representation of pointers, is now used only in
the implementations of CPointer and CTypedPointer.  It is no longer used
in the implementation of BoxedValueRecord.

There are two structures that implement C_MEMORY: CMemory and GMemory.
The former uses malloc, calloc and free whereas the latter uses
g_malloc, g_malloc0 and g_free.  CMemory is not currently used but is
provided for writing manual bindings to libraries not based on GLib.
CMemory should not be used for libraries based on GLib because ownership
is transferred on the expectation that allocation is done by g_malloc
(and variants) and freeing is done by g_free, and it should not be
assumed that these are interchangeable with malloc and free.  Therefore
the C interface type functors are instantiated with GMemory to derive
the C interface structures for use with GLib.  These structure names
have the prefix 'G' already except for FileDescType which is renamed to
GFileDescType and FileDesc which is renamed to GFileDesc.