Rename types that represent pointers to avoid misleading names
The type that represents a pointer has a type parameter in order to
distinguish between optional pointers, where null represents no value,
and non-optional pointers.  Prior to this change the type of a pointer
was as follows:

  - `notnull p` represents a non-optional pointer
  - `unit p`    represents an optional pointer

The phantom type `notnull` was misleading because a non-optional pointer
can legitimately have the value null provided it is never dereferenced.
This can occur in practice for an array of size zero: malloc may (and
does) return null if zero bytes are requested.  The pointer from malloc
is non-optional but it is misleading to describe it as `notnull`.  Also,
pointer arithmetic on a non-optional pointer could produce a null value
so again it is misleading to describe such values as `notnull` in the
type name.

Furthermore, `unit` is used to qualify a pointer as optional but it is
not a descriptive type name and its use may lead to confusion.

To address these issues, `notnull` is renamed to `non_opt` and a new
phantom type `opt` is introduced.  After this change, the type of a
pointer is as follows:

  - `non_opt p` represents a non-optional pointer
  - `opt p`     represents an optional pointer

This change to the phantom types that qualify a pointer type is
propagated through all C interface modules.