Move GLib-specific modules out of FFI directory
The GLib-specific modules in the FFI directory provided the predefined
GLib types assumed by GIR files and functors for constructing the GLib
types defined in GIR files.  These modules are moved to a new source
directory 'gir'.  The modules that introduce predefined GLib types,
namely the integer types and Utf8, are made visible from the GLib
namespace.  This allows users to write e.g. `GInt.null` with dependence
only on the GLib namespace.