Prefer zero-terminated type for a C array passed from a function
A C array function argument or return vaue may be zero-terminated and
have a size parameter.  If such a C array is passed from a function to
the caller (via either an 'out' parameter or the return value) it now
has the abstract type for a zero-terminated C array.  Although the size
parameter is not required, a null reference cannot be passed to the
function because the high-level FFI for MLton does not support a null
reference for a scalar type.

If a C array passed from the caller to a function (via either an 'in' or
'inout' parameter) is required to be zero-terminated and have a separate
size parameter, it still has the abstract type for an array with a
separate size parameter.  However, there do not appear to be any
examples of this, which is not surprising because it would place an
unnecessary burden on the caller.  Instead, the idiomatic approach to
allowing a C array passed to a function to be either zero-terminated or
use a size parameter is to allow the size parameter to be -1 for a
zero-terminated array.  The introspection annotations do not appear to
capture whether an array size parameter can be -1 for a null-terminated
array, so this cannot be exploited by a language binding.

This also fixes the function codegen design document in relation to the
handling of array lengths.