Rename function `sub` to `get` in sequence and C array modules
In the signatures SEQUENCE, MONO_SEQUENCE, C_ARRAY_TYPE and C_ARRAY,
the function `sub` is renamed to `get` and a new function `sub` is
introduced that is an uncurried variant of `get`.  This is done for
consistency with the signature VECTOR in the Basis Library.  The
signature C_ARRAY_TYPE already has a function called `get` so a
structure ElemSequence is introduced that contains the sequence
functions.