Support parameter ownership transfer in additional cases
For an 'in' parameter, support is added for transfer modes other than
'none'.

For a non-pointer 'inout' parameter with a struct type, support is added
for transfer mode 'everything'.  It does not make sense for such a
parameter to have transfer mode 'nothing' because the function would not
be allowed to modify the parameter, so an error is raised for this.

An 'out' parameter that is caller-allocated cannot transfer ownership so
an error is raised if the transfer mode is not 'nothing'.