Support caller-allocates 'out' C array parameters
An 'out' C array parameter that is caller-allocated is supported only if
the annotations specify a length for the array that is a fixed size or a
length parameter with direction 'in'.  In the codegen, such a length
parameter is treated as an ordinary, non-length, parameter in certain
respects in order to expose it to the caller.  In the C array interface,
the function `withNewPtr` is added to the high-level FFI.

The codegen for a parameter or return value with type tag UTF8 or
FILENAME is also updated, to perform checks consistently with ARRAY.
However, an 'out' parameter with type tag UTF8 or FILENAME cannot be
caller-allocated because the functions in the girepository interface
that provide the fixed size or length parameter of an array work only
for the type tag ARRAY.

The functions g_input_stream_read* and g_socket_receive* should treat
the parameter `buffer` as a caller-allocates 'out' parameter.  For
details, see:

  https://gitlab.gnome.org/GNOME/glib/-/issues/1258
  https://gitlab.gnome.org/GNOME/glib/-/issues/2176

As these changes are upstream, the introspection annotations are
overridden.

In the codegen design document for functions, out of date statements
relating to the value returned by `GIRepository.ArgInfo.mayBeNull` and
`GIRepository.CallableInfo.mayReturnNull` are removed.